[wired-combo] Handle large wired-item lists.
If the list of <wired-item>s is very large, the <wired-card> element
used to display then is tall causing the items at the bottom to be
below the window and unselectable.  This change limits the maximum
size of the card to 40% of window height and makes the card scollable
in order to reach the lower items.